refactor: extract config state into src/config.ts (break hooks<->index cycle)
`hooks.ts` imported `configureAuthKit` from `./index.js`; `index.ts` imported `createAuthKitHandle` from `./hooks.js`. The cycle resolved at runtime only because `configureAuthKit` was referenced inside a callback (ESM hoisting), but any future move to module scope in `hooks.ts` would TDZ-fail at import.

Move the shared state (`authKitInstance`, `configuredConfig`), validation, and `configureAuthKit` / `getAuthKitInstance` to `src/config.ts`. Both `index.ts` and `hooks.ts` now import from `config.ts`; `index.ts` re-exports `configureAuthKit` so the public surface is unchanged.
